pyproject.toml
[project]
name = "mlinks"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "aerich[toml]>=0.9.0",
    "celery>=5.5.2",
    "fastapi[standard]>=0.115.12",
    "flower>=2.0.1",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "redis[hiredis]>=6.1.0",
    "tortoise-orm[asyncpg]>=0.25.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "ruff>=0.11.10",
]

[tool.aerich]
tortoise_orm = "db.database.TORTOISE_ORM_CONFIG"
location = "./src/db/migrations" 
src_folder = "./src" 
app = "models" 

[tool.ruff]
target-version = "py313"
line-length = 120
src = [
    "src",
]
exclude = [
    "migrations",
    "__pycache__",
]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "UP",
    "B",
    "C4",
    "SIM",
    "ANN",
    "TRY",
    "RUF",
    "TID",
    "PT",
]
ignore = [
    "E501",
    "UP043",
    "RUF002",
    "ANN401",
]
fixable = ["ALL"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["src"]
combine-as-imports = true


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

